Move pkg/api/events to only user pkg/printers/internalversion
